requirements.txt
torch==2.3.1
torchvision==0.18.1

addict==2.4.0
ansi2html==1.8.0
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==3.0.0
coacd==1.0.0
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.18.1
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.1
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
gymnasium==0.29.1
h5py==3.10.0
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipython==8.12.3
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
mani-skill2==0.5.3
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
MeshLabXML==2018.3
mplib==0.0.9
nbformat==5.7.0
nest-asyncio==1.5.8
numpy==1.23.5
open3d==0.17.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.5
pyglet==1.5.28
Pygments==2.16.1
pymesh==1.0.2
pymeshlab==2022.2.post4
pyparsing==3.1.1
pyquaternion==0.9.9
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rpds-py==0.10.6
Rtree==1.1.0
sapien==2.2.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tomli==2.0.1
toppra==0.6.0
tqdm==4.66.1
traitlets==5.13.0
transformations==2022.9.26
transforms3d==0.4.1
trimesh==4.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
widgetsnbextension==4.0.9
zipp==3.17.0
torchmetrics==0.7.0
supervision==0.22.0

# RGBManip
tensorboard==2.12.0
ujson==5.7.0
ipdb
hydra-core==1.3.2
gym==0.26.2
